api cleanup, basic structure
